chore: remove old scaffold code
