+++ environment.yaml
@@ -2,10 +2,16 @@ name: rsas_env
 channels:
 - conda-forge
 dependencies:
+- annotated-types=0.7.0=pyhd8ed1ab_0
+- anyio=4.4.0=pyhd8ed1ab_0
 - bzip2=1.0.8=h99b78c6_7
 - ca-certificates=2024.7.4=hf0a4a13_0
 - cairo=1.18.0=hb4a6bf7_3
+- certifi=2024.8.30=pyhd8ed1ab_0
+- colorama=0.4.6=pyhd8ed1ab_0
+- distro=1.9.0=pyhd8ed1ab_0
 - double-conversion=3.3.0=h13dd4ca_0
+- exceptiongroup=1.2.2=pyhd8ed1ab_0
 - expat=2.6.2=hebf3989_0
 - font-ttf-dejavu-sans-mono=2.37=hab24e00_0
 - font-ttf-inconsolata=3.000=h77eed37_0
@@ -16,8 +22,16 @@ dependencies:
 - fonts-conda-forge=1=0
 - freetype=2.12.1=hadb7bae_2
 - graphite2=1.3.13=hebf3989_1003
+- h11=0.14.0=pyhd8ed1ab_0
+- h2=4.1.0=pyhd8ed1ab_0
 - harfbuzz=9.0.0=h997cde5_1
+- hpack=4.0.0=pyh9f0ad1d_0
+- httpcore=1.0.5=pyhd8ed1ab_0
+- httpx=0.27.2=pyhd8ed1ab_0
+- hyperframe=6.0.1=pyhd8ed1ab_0
 - icu=75.1=hfee45f7_0
+- idna=3.8=pyhd8ed1ab_0
+- jiter=0.5.0=py312he431725_1
 - krb5=1.21.3=h237132a_0
 - lerc=4.0.0=h9a09cb3_0
 - libblas=3.9.0=23_osxarm64_openblas
@@ -52,11 +66,14 @@ dependencies:
 - mysql-libs=8.3.0=h0e80b4a_5
 - ncurses=6.5=hb89a1cb_0
 - numpy=2.0.1=py312hb544834_0
+- openai=1.44.1=pyhd8ed1ab_0
 - openssl=3.3.1=hfb2fe0b_2
 - pandas=2.2.2=py312h8ae5369_1
 - pcre2=10.44=h297a79d_2
 - pip=24.2=pyhd8ed1ab_0
 - pixman=0.43.4=hebf3989_0
+- pydantic=2.9.1=pyhd8ed1ab_0
+- pydantic-core=2.23.3=py312he431725_0
 - pyside6=6.7.2=py312h7412757_2
 - python=3.12.5=h30c5eda_0_cpython
 - python-dateutil=2.9.0=pyhd8ed1ab_0
@@ -68,7 +85,11 @@ dependencies:
 - ruff=0.6.1=py312h3402d49_0
 - setuptools=72.1.0=pyhd8ed1ab_0
 - six=1.16.0=pyh6c4a22f_0
+- sniffio=1.3.1=pyhd8ed1ab_0
 - tk=8.6.13=h5083fa2_1
+- tqdm=4.66.5=pyhd8ed1ab_0
+- typing-extensions=4.12.2=hd8ed1ab_0
+- typing_extensions=4.12.2=pyha770c72_0
 - tzdata=2024a=h0c530f3_0
 - wheel=0.44.0=pyhd8ed1ab_0
 - xz=5.2.6=h57fd34a_0
